requirements.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu
sentence-transformers
langchain
langchain-community
huggingface_hub

